Add: Dynamic module loading for subparsers
