+++ matrixctl/addons/check.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Use this module to add the ``deploy`` subcommand to ``matrixctl``."""
+"""Use this module to add the ``check`` subcommand to ``matrixctl``."""
 
 from __future__ import annotations
 
@@ -25,8 +25,10 @@ from argparse import ArgumentParser
 from argparse import Namespace
 from argparse import _SubParsersAction as SubParsersAction
 
-from .handlers.ansible import ansible_run
-from .handlers.yaml import YAML
+from argparse_addon_manager.addon_manager import AddonManager
+
+from matrixctl.handlers.ansible import ansible_run
+from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
@@ -36,8 +38,9 @@ __email__: str = "Michael@MichaelSasser.org"
 logger = logging.getLogger(__name__)
 
 
-def subparser_deploy(subparsers: SubParsersAction) -> None:
-    """Create a subparser for the ``matrixctl deploy`` command.
+@AddonManager.add_subparser
+def subparser_check(subparsers: SubParsersAction) -> None:
+    """Create a subparser for the ``matrixctl check`` command.
 
     Parameters
     ----------
@@ -51,20 +54,13 @@ def subparser_deploy(subparsers: SubParsersAction) -> None:
 
     """
     parser: ArgumentParser = subparsers.add_parser(
-        "deploy", help="Provision and deploy"
+        "check", help="Checks the deployment with ansible"
     )
-
-    parser.add_argument(  # Done with tags / Does not use matrixctl.start !
-        "-s",
-        "--start",
-        action="store_true",
-        help="Start/Restart after the deployment",
-    )
-    parser.set_defaults(func=deploy)
+    parser.set_defaults(func=check)
 
 
-def deploy(arg: Namespace, yaml: YAML) -> int:
-    """Deploy the ansible playbook.
+def check(_: Namespace, yaml: YAML) -> int:
+    """Check the deployment with andible.
 
     Parameters
     ----------
@@ -79,11 +75,9 @@ def deploy(arg: Namespace, yaml: YAML) -> int:
         Non-zero value indicates error code, or zero on success.
 
     """
-    ansible_run(
-        playbook=yaml.get("ansible", "playbook"),
-        tags="setup-all,start" if arg.start else "setup-all",
-    )
+    logger.debug("check")
 
+    ansible_run(playbook=yaml.get("ansible", "playbook"), tags="check")
     return 0
