+++ matrixctl/addons/get_event.py
@@ -28,9 +28,11 @@ from argparse import Namespace
 from argparse import _SubParsersAction as SubParsersAction
 from base64 import b64encode
 
-from .handlers.ssh import SSH
-from .handlers.ssh import SSHResponse
-from .handlers.yaml import YAML
+from argparse_addon_manager.addon_manager import AddonManager
+
+from matrixctl.handlers.ssh import SSH
+from matrixctl.handlers.ssh import SSHResponse
+from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
@@ -42,6 +44,7 @@ logger = logging.getLogger(__name__)
 JID_EXT: str = "matrix-jitsi-web"
 
 
+@AddonManager.add_subparser
 def subparser_get_event(subparsers: SubParsersAction) -> None:
     """Create a subparser for the ``matrixctl get-event`` command.
