+++ matrixctl/addons/start.py
@@ -15,24 +15,32 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Use this module to add the ``update`` subcommand to ``matrixctl``."""
+"""Use this module to add the ``(re)start`` subcommand to ``matrixctl``."""
 
 from __future__ import annotations
+
+import logging
 
 from argparse import ArgumentParser
 from argparse import Namespace
 from argparse import _SubParsersAction as SubParsersAction
 
-from .handlers.vcs import VCS
-from .handlers.yaml import YAML
+from argparse_addon_manager.addon_manager import AddonManager
+
+from matrixctl.handlers.ansible import ansible_run
+from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
 __email__: str = "Michael@MichaelSasser.org"
 
 
-def subparser_update(subparsers: SubParsersAction) -> None:
-    """Create a subparser for the ``matrixctl update`` command.
+logger = logging.getLogger(__name__)
+
+
+@AddonManager.add_subparser
+def subparser_start(subparsers: SubParsersAction) -> None:
+    """Create a subparser for the ``matrixctl start`` command.
 
     Parameters
     ----------
@@ -45,13 +53,40 @@ def subparser_update(subparsers: SubParsersAction) -> None:
 
     """
     parser: ArgumentParser = subparsers.add_parser(
-        "update", help="Updates the ansible repo"
+        "start", help="Starts all OCI containers"
     )
-    parser.set_defaults(func=update)
+    parser.set_defaults(func=start)
 
 
-def update(_: Namespace, yaml: YAML) -> int:
-    """Update the synapse playbook with git.
+def subparser_restart(subparsers: SubParsersAction) -> None:
+    """Create a subparser for the ``matrixctl restart`` command.
+
+    Notes
+    -----
+    This is a alias for ``matrixctl start``
+
+    See Also
+    --------
+    matrixctl.start.subparser_start : Subparser for ``matrixctl start``.
+
+    Parameters
+    ----------
+    subparsers : argparse._SubParsersAction
+        The object which is returned by ``parser.add_subparsers()``.
+
+    Returns
+    -------
+    None
+
+    """
+    parser: ArgumentParser = subparsers.add_parser(
+        "restart", help="Restarts all OCI containers (alias for start)"
+    )
+    parser.set_defaults(func=start)  # Keep it "start"
+
+
+def start(_: Namespace, yaml: YAML) -> int:
+    """Start/Restart the OCI containers.
 
     Parameters
     ----------
@@ -66,9 +101,7 @@ def update(_: Namespace, yaml: YAML) -> int:
         Non-zero value indicates error code, or zero on success.
 
     """
-    git: VCS = VCS(yaml.get("synapse", "playbook"))
-    git.pull()
-
+    ansible_run(yaml.get("ansible", "playbook"), tags="start")
     return 0
