+++ matrixctl/addons/stop.py
@@ -15,37 +15,32 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Use this module to add the ``deluser`` subcommand to ``matrixctl``."""
+"""Use this module to add the ``stop`` subcommand to ``matrixctl``."""
 
 from __future__ import annotations
-
-import logging
 
 from argparse import ArgumentParser
 from argparse import Namespace
 from argparse import _SubParsersAction as SubParsersAction
 
-from .errors import InternalResponseError
-from .handlers.api import RequestBuilder
-from .handlers.api import request
-from .handlers.yaml import YAML
+from argparse_addon_manager.addon_manager import AddonManager
+
+from matrixctl.handlers.ansible import ansible_run
+from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
 __email__: str = "Michael@MichaelSasser.org"
 
 
-logger = logging.getLogger(__name__)
-
-
-def subparser_deluser(subparsers: SubParsersAction) -> None:
-    """Create a subparser for the ``matrixctl deluser`` command.
+@AddonManager.add_subparser
+def subparser_stop(subparsers: SubParsersAction) -> None:
+    """Create a subparser for the ``matrixctl stop`` command.
 
     Parameters
     ----------
     subparsers : argparse._SubParsersAction
-        The object which is returned by
-        ``parser.add_subparsers()``.
+        The object which is returned by ``parser.add_subparsers()``.
 
     Returns
     -------
@@ -53,19 +48,18 @@ def subparser_deluser(subparsers: SubParsersAction) -> None:
 
     """
     parser: ArgumentParser = subparsers.add_parser(
-        "deluser", help="Deletes a user"
+        "stop", help="Stops all OCI containers"
     )
-    parser.add_argument("user", help="The username to delete")
-    parser.set_defaults(func=deluser)
+    parser.set_defaults(func=stop)
 
 
-def deluser(arg: Namespace, yaml: YAML) -> int:
-    """Delete a user from the the matrix instance.
+def stop(_: Namespace, yaml: YAML) -> int:
+    """Stop the OCI containers.
 
     Parameters
     ----------
     arg : argparse.Namespace
-        The ``Namespace`` object of argparse's ``parse_args()``
+        The ``Namespace`` object of argparse's ``parse_args()``.
     yaml : matrixctl.handlers.yaml.YAML
         The configuration file handler.
 
@@ -75,19 +69,7 @@ def deluser(arg: Namespace, yaml: YAML) -> int:
         Non-zero value indicates error code, or zero on success.
 
     """
-    req: RequestBuilder = RequestBuilder(
-        token=yaml.get("api", "token"),
-        domain=yaml.get("api", "domain"),
-        path=f"deactivate/@{arg.user}:{yaml.get('api','domain')}",
-        api_version="v1",
-        method="POST",
-        data={"erase": True},
-    )
-    try:
-        request(req)
-    except InternalResponseError:
-        logger.error("The user was not deleted.")
-
+    ansible_run(yaml.get("ansible", "playbook"), tags="stop")
     return 0
