+++ matrixctl/addons/upload.py
@@ -27,20 +27,22 @@ from argparse import _SubParsersAction as SubParsersAction
 from mimetypes import MimeTypes
 from pathlib import Path
 
-from .errors import InternalResponseError
-from .handlers.api import RequestBuilder
-from .handlers.api import request
-from .handlers.yaml import YAML
-from .typehints import JsonDict
+from argparse_addon_manager.addon_manager import AddonManager
+
+from matrixctl.errors import InternalResponseError
+from matrixctl.handlers.api import RequestBuilder
+from matrixctl.handlers.api import request
+from matrixctl.handlers.yaml import YAML
+from matrixctl.typehints import JsonDict
 
 
 __author__: str = "Michael Sasser"
 __email__: str = "Michael@MichaelSasser.org"
 
-
 logger = logging.getLogger(__name__)
 
 
+@AddonManager.add_subparser
 def subparser_upload(subparsers: SubParsersAction) -> None:
     """Create a subparser for the ``matrixctl upload`` command.
